Moved fs watcher and improved logging
