+++ internal/fs/watcher_test.go
